refactor: Double -> BigDecimal へ変更

+++ app/src/main/java/com/example/yosuke/calculator/model/entity/CalcEntity.kt
@@ -1,6 +1,8 @@
 package com.example.yosuke.calculator.model.entity
 
-data class CalcEntity(val number: Double, var operator: Operators) {
+import java.math.BigDecimal
+
+data class CalcEntity(val number: BigDecimal, var operator: Operators) {
     override fun toString(): String {
         return "$number ${operator.toStr()}"
     }
